chore(card): move sub components to dir
