+++ packages/card/src/components/tabPane.tsx
@@ -4,7 +4,7 @@ import { TabPaneProps } from 'antd/lib/tabs';
 import classNames from 'classnames';
 
 import './style/index.less';
-import ProCard, { ProCardProps } from './index';
+import ProCard, { ProCardProps } from '../index';
 
 export interface ProCardTabPaneProps extends TabPaneProps, ProCardProps {
   key?: string;
